Refactor Logger->log to avoid creating a POE::Session.
A major refactoring of Logger->log/P:C:L->log to remove the creation of
an ephemerous POE::Session that was used just for sending the message
to our logger session.
The code change is low, but it may have big impact. Probably not because
the _start event is called synchronously in POE::Session->create().
Anyway, our extensive test suite passes.